ibc-types: use `ibc-proto@0.52`, prep `v0.16` (#97)
This patch bumps the recently released `ibc-proto@v0.52.0` and prepares
a new release of all crates that have it as a direct or transient
dependency:

+ ibc-types@v0.16.0
+ ibc-types-core-channel@v0.16.0
+ ibc-types-core-client@v0.16.0
+ ibc-types-core-commitment@v0.16.0
+ ibc-types-core-connection@v0.16.0
+ ibc-types-lightclients-tendermint@v0.16.0
+ ibc-types-path@v0.16.0

Because ibc-proto now has a msrv of rust 1.75.0 because of its
dependency on tonic, I have set `rust-version = "1.75.0"` in all the
aforementioned crates.

Note that this patch currently breaks with the convention of bumping all
crate versions in the workspace in lock-step: some of the crates are not
affected by the changes and so I have opted to not change them. If it is
desired to have a lockstep release I can also do that.